# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ModBridge API",
    "description": "Modbus TCP Proxy Manager API. All endpoints except the health, readiness, metrics, login, setup and recovery routes require an authenticated session; state-changing requests additionally require a CSRF token.",
    "version": "dev"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/devices": {
      "get": {
        "summary": "List or rename tracked devices",
        "description": "",
        "operationId": "getDevices",
        "responses": {
          "200": {
            "description": "Success",
            "content": null
          },
          "401": {
            "description": "Not authenticated",
            "content": null
          },
          "403": {
            "description": "Not permitted",
            "content": null
          }
        }
      },
      "put": {
        "summary": "List or rename tracked devices",
        "description": "",
        "operationId": "putDevices",
        "responses": {
          "200": {
            "description": "Success",
            "content": null
          },
          "401": {
            "description": "Not authenticated",
            "content": null
          },
          "403": {
            "description": "Not permitted",
            "content": null
          }
        }
      }
    },
...
    "/api/logs": {
      "get": {
        "summary": "Read log entries",
        "description": "",
        "operationId": "getLogs",
        "responses": {
          "200": {
            "description": "Success",
            "content": null
          },
          "401": {
            "description": "Not authenticated",
            "content": null
          },
          "403": {
            "description": "Not permitted",
            "content": null
          }
        }
      }
    },
...
    "/api/proxies": {
      "delete": {
        "summary": "List, create, update or delete proxies",
        "description": "",
        "operationId": "deleteProxies",
        "responses": {
          "200": {
            "description": "Success",
            "content": null
          },
          "401": {
            "description": "Not authenticated",
            "content": null
          },
          "403": {
            "description": "Not permitted",
            "content": null
          }
        }
      },
      "get": {
        "summary": "List, create, update or delete proxies",
        "description": "",
        "operationId": "getProxies",
        "responses": {
          "200": {
            "description": "Success",
            "content": null
          },
          "401": {
            "description": "Not authenticated",
            "content": null
          },
          "403": {
            "description": "Not permitted",
            "content": null
          }
        }
      },
      "post": {
        "summary": "List, create, update or delete proxies",
        "description": "",
        "operationId": "postProxies",
        "responses": {
          "200": {
            "description": "Success",
            "content": null
          },
          "401": {
            "description": "Not authenticated",
            "content": null
          },
          "403": {
            "description": "Not permitted",
            "content": null
          }
        }
      },
      "put": {
        "summary": "List, create, update or delete proxies",
        "description": "",
        "operationId": "putProxies",
        "responses": {
          "200": {
            "description": "Success",
            "content": null
          },
          "401": {
            "description": "Not authenticated",
            "content": null
          },
          "403": {
            "description": "Not permitted",
            "content": null
          }
        }
      }
    },
...
  },
...
}